--book work started
In the new "book" branch, the database auto-updates to a version in
which the "in_trash" column of the "note" table is replaced by "book".
(This requires inversion, since book 0 will henceforth mean the trash.)
The new version seems to reproduce the existing version, but I plan to
use it a while before merging into the "develop" branch.

As a useful coding side-effect, I now know how to rename the column of a
table in sqlite3, which is trickier than I would have thought.